requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
confluent-kafka
pandas
numpy
scikit-learn
tensorflow
joblib
